+++ src/pages/export/ExportPage.vue
@@ -15,6 +15,7 @@ import { parseAddressString } from 'src/lib/function-interface-utils';
 import { contractManager, indexerApi } from 'src/boot/telosApi';
 import { ZERO_ADDRESSES } from 'src/lib/utils';
 import { addEmptyContractToCache } from 'src/lib/contract-utils';
+import { formatTimestamp } from 'src/lib/date-utils';
 
 import { EvmTransaction } from 'src/antelope/types/EvmTransaction';
 
@@ -132,19 +133,6 @@ async function download() {
         }
 
         return value;
-    }
-
-    function formatTimestamp(unixTimestampMs: number) {
-        const date = new Date(unixTimestampMs);
-
-        const year = date.getFullYear();
-        const month = String(date.getMonth() + 1).padStart(2, '0'); // Month is 0-indexed
-        const day = String(date.getDate()).padStart(2, '0');
-        const hours = String(date.getHours()).padStart(2, '0');
-        const minutes = String(date.getMinutes()).padStart(2, '0');
-        const seconds = String(date.getSeconds()).padStart(2, '0');
-
-        return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
     }
 
     if (typeSelectModel.value.value === EXPORT_DOWNLOAD_TYPES.transactions) {
